+++ bundle/src/test/java/com/adobe/acs/commons/wcm/notifications/impl/SystemNotificationsImplTest.java
@@ -27,26 +27,29 @@ import static org.hamcrest.text.StringContainsInOrder.stringContainsInOrder;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertThat;
 
-import io.wcm.testing.mock.aem.junit.AemContext;
 import java.util.Arrays;
 import java.util.Calendar;
 import java.util.Collections;
 import java.util.Date;
+
 import javax.servlet.Filter;
 import javax.servlet.FilterChain;
 import javax.servlet.http.Cookie;
+
 import org.apache.commons.codec.digest.DigestUtils;
-import org.apache.sling.api.SlingConstants;
 import org.apache.sling.api.resource.PersistenceException;
 import org.apache.sling.api.resource.ResourceResolver;
+import org.apache.sling.api.resource.observation.ResourceChange;
+import org.apache.sling.api.resource.observation.ResourceChange.ChangeType;
+import org.apache.sling.api.resource.observation.ResourceChangeListener;
 import org.apache.sling.settings.SlingSettingsService;
 import org.apache.sling.testing.mock.sling.ResourceResolverType;
 import org.apache.sling.testing.mock.sling.services.MockSlingSettingService;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
-import org.osgi.service.event.Event;
-import org.osgi.service.event.EventHandler;
+
+import io.wcm.testing.mock.aem.junit.AemContext;
 
 public class SystemNotificationsImplTest {
 
@@ -68,7 +71,7 @@ public class SystemNotificationsImplTest {
     @Test
     public void testOnPublish() {
         aemContext.registerInjectActivateService(notifications);
-        assertEquals(0, aemContext.getServices(EventHandler.class, null).length);
+        assertEquals(0, aemContext.getServices(ResourceChangeListener.class, null).length);
         assertEquals(0, aemContext.getServices(Filter.class, null).length);
     }
 
@@ -76,7 +79,7 @@ public class SystemNotificationsImplTest {
     public void testOnAuthorNoNotifications() {
         setAuthorRunmode();
         aemContext.registerInjectActivateService(notifications);
-        assertEquals(1, aemContext.getServices(EventHandler.class, null).length);
+        assertEquals(1, aemContext.getServices(ResourceChangeListener.class, null).length);
         assertEquals(0, aemContext.getServices(Filter.class, null).length);
     }
 
@@ -88,7 +91,7 @@ public class SystemNotificationsImplTest {
                 .resource("jcr:content", JCR_PRIMARYTYPE, NT_UNSTRUCTURED);
         commit();
         aemContext.registerInjectActivateService(notifications);
-        assertEquals(1, aemContext.getServices(EventHandler.class, null).length);
+        assertEquals(1, aemContext.getServices(ResourceChangeListener.class, null).length);
         assertEquals(0, aemContext.getServices(Filter.class, null).length);
 
         aemContext.create().resource("/etc/acs-commons/notifications/first", JCR_PRIMARYTYPE, NT_PAGE);
@@ -98,7 +101,7 @@ public class SystemNotificationsImplTest {
 
         notifications.deactivate(aemContext.componentContext());
         assertEquals(0, aemContext.getServices(Filter.class, null).length);
-        assertEquals(0, aemContext.getServices(EventHandler.class, null).length);
+        assertEquals(0, aemContext.getServices(ResourceChangeListener.class, null).length);
     }
 
     @Test
@@ -106,7 +109,7 @@ public class SystemNotificationsImplTest {
         setAuthorRunmode();
         createEnabledNotification();
         aemContext.registerInjectActivateService(notifications);
-        assertEquals(1, aemContext.getServices(EventHandler.class, null).length);
+        assertEquals(1, aemContext.getServices(ResourceChangeListener.class, null).length);
         assertEquals(1, aemContext.getServices(Filter.class, null).length);
 
         // then remove the resource
@@ -117,7 +120,7 @@ public class SystemNotificationsImplTest {
 
         notifications.deactivate(aemContext.componentContext());
         assertEquals(0, aemContext.getServices(Filter.class, null).length);
-        assertEquals(0, aemContext.getServices(EventHandler.class, null).length);
+        assertEquals(0, aemContext.getServices(ResourceChangeListener.class, null).length);
     }
 
     @Test
@@ -129,7 +132,7 @@ public class SystemNotificationsImplTest {
 
         createEnabledNotification("/etc/acs-commons/notifications/subfolder/enabled");
         aemContext.registerInjectActivateService(notifications);
-        assertEquals(1, aemContext.getServices(EventHandler.class, null).length);
+        assertEquals(1, aemContext.getServices(ResourceChangeListener.class, null).length);
         assertEquals(1, aemContext.getServices(Filter.class, null).length);
 
         // then remove the resource
@@ -140,7 +143,7 @@ public class SystemNotificationsImplTest {
 
         notifications.deactivate(aemContext.componentContext());
         assertEquals(0, aemContext.getServices(Filter.class, null).length);
-        assertEquals(0, aemContext.getServices(EventHandler.class, null).length);
+        assertEquals(0, aemContext.getServices(ResourceChangeListener.class, null).length);
     }
 
     @Test
@@ -227,7 +230,7 @@ public class SystemNotificationsImplTest {
     }
 
     private void sendEvent() {
-        notifications.handleEvent(new Event("dummy", Collections.singletonMap(SlingConstants.PROPERTY_PATH, "/etc/acs-commons/notifications")));
+        notifications.onChange(Collections.singletonList(new ResourceChange(ChangeType.ADDED, "/etc/acs-commons/notifications", false, null, null, null)));
     }
 
     private void delete(String path) throws PersistenceException {